Relocate test utils
