+++ src/forms/utils/createCoverPage.ts
@@ -1,0 +1,151 @@
+import { smartquotes } from "~/utils/smartquotes";
+import { loadPdfLib } from "./loadPdfLib";
+
+/**
+ * Create a cover page PDF for a document packet.
+ */
+export async function createCoverPage({
+  title,
+  instructions,
+  documents,
+}: {
+  title: string;
+  instructions: string[];
+  documents: Array<{ title: string; code?: string }>;
+}): Promise<Uint8Array> {
+  const {
+    PDFDocument,
+    StandardFonts,
+    pushGraphicsState,
+    popGraphicsState,
+    setCharacterSpacing,
+  } = await loadPdfLib();
+
+  const pdfDoc = await PDFDocument.create();
+  const page = pdfDoc.addPage([612, 792]); // Standard US Letter size
+  const helveticaBold = await pdfDoc.embedFont(StandardFonts.HelveticaBold);
+  const helvetica = await pdfDoc.embedFont(StandardFonts.Helvetica);
+  const { height, width } = page.getSize();
+
+  const margin = 50;
+  const contentWidth = width - margin * 2;
+
+  function drawList({
+    title,
+    items,
+    y,
+  }: {
+    title: string;
+    items: string[];
+    y: number;
+  }) {
+    const listOffset = 150;
+
+    page.drawText(smartquotes(title), {
+      x: margin,
+      y: y,
+      size: 16,
+      font: helveticaBold,
+    });
+
+    const listText = items
+      .filter((item) => item.trim().length > 0)
+      .map((item) => smartquotes(item))
+      .join("\n");
+
+    page.drawText(listText, {
+      x: margin + listOffset,
+      y: y,
+      size: 13,
+      font: helvetica,
+      maxWidth: contentWidth - listOffset,
+      wordBreaks: [" "],
+      lineHeight: 21,
+    });
+  }
+
+  page.drawText("Name Change Packet", {
+    x: margin,
+    y: height - 65,
+    size: 14,
+    font: helveticaBold,
+  });
+
+  page.drawLine({
+    start: { x: margin, y: height - 80 },
+    end: { x: width - margin, y: height - 80 },
+    thickness: 2.5,
+  });
+
+  page.pushOperators(pushGraphicsState(), setCharacterSpacing(-1));
+  page.drawText(smartquotes(title), {
+    x: margin - 1.5, // Optical fix
+    y: height - 130,
+    size: 44,
+    lineHeight: 48,
+    maxWidth: contentWidth,
+    wordBreaks: [" "],
+    font: helveticaBold,
+  });
+  page.pushOperators(popGraphicsState());
+
+  drawList({
+    title: "Packet Includes",
+    items: documents.map((doc) =>
+      doc.code ? `${doc.title} (${doc.code})` : doc.title,
+    ),
+    y: height - 270,
+  });
+
+  drawList({
+    title: "Instructions",
+    items: instructions,
+    y: height - 420,
+  });
+
+  try {
+    const logoResponse = await fetch("/forms/pdf-cover-logo.png");
+    const logoBytes = await logoResponse.arrayBuffer();
+    const logoImage = await pdfDoc.embedPng(new Uint8Array(logoBytes));
+
+    page.drawImage(logoImage, {
+      x: margin,
+      y: 105,
+      height: 16,
+      width: 83,
+    });
+  } catch (error) {
+    console.warn("Failed to load or draw logo:", error);
+  }
+
+  const currentDate = new Date().toLocaleDateString("en-US", {
+    month: "long",
+    day: "numeric",
+    year: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+  const generatedOn = `Document generated by https://app.namesake.fyi on ${currentDate}.`;
+  page.drawText(smartquotes(generatedOn), {
+    x: margin,
+    y: 80,
+    size: 9,
+    lineHeight: 11,
+    font: helvetica,
+  });
+
+  const disclaimer =
+    "Disclaimer: The information provided by Namesake is for general informational purposes and does not constitute legal advice. Use of namesake.fyi does not create an attorney-client relationship between you and Namesake.";
+  page.drawText(smartquotes(disclaimer), {
+    x: margin,
+    y: 60,
+    size: 9,
+    font: helvetica,
+    maxWidth: contentWidth,
+    wordBreaks: [" "],
+    lineHeight: 11,
+  });
+
+  return await pdfDoc.save();
+}
+
