+++ src/forms/utils/definePdf.ts
@@ -1,0 +1,10 @@
+import type { PDFDefinition } from "~/constants/forms";
+
+/**
+ * Define a PDF form.
+ * @returns A PDF definition.
+ */
+export function definePdf(pdf: PDFDefinition): PDFDefinition {
+  return pdf;
+}
+
